+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This version includes many fixes and improvements mentioned in the survey. Thank
 * [feat]: Adds button to reset the "hide indices regex" setting to default
 * [feat]: When deleting an index the name of the index is now included in the confirm message
 * [feat]: Adds "Copy content" button to editors
+* [feat]: Can download search response as json
 * [feat]: Can download rest query response as json
 * [feat]: Cleanup code editor settings
 * [feat]: Simplify footer, move link to reset settings to settings page